pd: split wallet code into modules
